tests/wpt/web-platform-tests/trusted-types/HTMLLinkElement-href.tentative.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/helper.sub.js"></script>
<body>
<script>
  //helper function for the tests
  function testHref(str, url) {
    var link = document.createElement('link');
    link.href = url;
    assert_equals(link.href, str);
  }

  test(t => {
    testHref(URLS.safe, TrustedURL.create(URLS.safe));
  }, "link.href = URLS.safe, TrustedURL.create");

  test(t => {
    testHref(URLS.safe, TrustedURL.unsafelyCreate(URLS.safe));
  }, "link.href = URLS.safe, TrustedURL.unsafelyCreate");
</script>
